뷰 바인딩, 폴더링 리펙토링
