+++ app/src/main/java/com/seoultech/fooddeuk/intro/IntroViewPagerAdapter.java
@@ -1,4 +1,4 @@
-package com.seoultech.fooddeuk;
+package com.seoultech.fooddeuk.intro;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -9,6 +9,8 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.seoultech.fooddeuk.R;
 
 import org.jetbrains.annotations.NotNull;
