+++ app/src/main/res/layout/activity_intro_customer.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#ffffff"
-    tools:context=".IntroCustomerActivity">
+    tools:context=".intro.IntroCustomerActivity">
 
     <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/btnSkip"
