+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".SignUpActivity">
+    tools:context=".signUp.SignUpActivity">
 
     <Button
         android:id="@+id/tmp_sign_up_button"
