+++ app/src/main/res/layout/layout_guide_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#ffffff"
-    tools:context=".IntroCustomerActivity">
+    tools:context=".intro.IntroCustomerActivity">
 
     <TextView
         android:id="@+id/tvTitle"
